requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
filelock==3.8.0
huggingface-hub==0.11.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numpy
opencv-contrib-python==4.5.4.60
opencv-python==4.6.0.66
packaging==21.3
pillow
pipreqs==0.4.13
pyparsing==3.0.9
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.7.3
six
tabulate==0.9.0
tifffile==2021.11.2
timm==0.6.12
torch==1.10.1
torchaudio==0.10.1
torchvideotransforms @ git+https://github.com/hassony2/torch_videovision@785a810d1df2ccab6901fb98b38e358433e7b17a
torchvision==0.11.2
tqdm==4.64.1
typing_extensions
urllib3==1.26.13
yarg==0.1.9
zipp==3.10.0
